move diagnostic updater for stream frequencies into the RosSensor class.
